# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RegMail API",
        "description": "API for RegMail - Email Registration Service",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/api/email/submit": {
            "post": {
                "tags": [
                    "Email Submission"
                ],
                "summary": "Submit successful email registration",
                "description": "Submit details of successfully created email account with registration time tracking. Device fingerprint should be generated by client using: import hashlib, platform, uuid; system_info = platform.platform() + platform.machine() + str(uuid.getnode()); fingerprint = 'device_' + hashlib.sha256(system_info.encode()).hexdigest()[:12]",
                "operationId": "2081724902b9db23e89d49c21962ab75",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "proxy_info",
                                    "device_fingerprint"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "testuser123@gmail.com"
                                    },
                                    "recovery_email": {
                                        "description": "Recovery email for account recovery",
                                        "type": "string",
                                        "format": "email",
                                        "example": "recovery123@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "SecurePass123!"
                                    },
                                    "registration_time": {
                                        "description": "Time in seconds from start to successful registration (optional, defaults to 0)",
                                        "type": "integer",
                                        "example": 3600
                                    },
                                    "device_fingerprint": {
                                        "description": "Unique device identifier for tracking which device registered the email",
                                        "type": "string",
                                        "example": "device_abc123xyz"
                                    },
                                    "proxy_info": {
                                        "properties": {
                                            "ip": {
                                                "type": "string",
                                                "example": "192.168.1.100"
                                            },
                                            "port": {
                                                "type": "integer",
                                                "example": 8080
                                            },
                                            "username": {
                                                "type": "string",
                                                "example": "proxy_user"
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": "proxy_pass"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "metadata": {
                                        "properties": {
                                            "user_agent": {
                                                "type": "string",
                                                "example": "Mozilla/5.0..."
                                            },
                                            "creation_time": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-10-24T21:30:00Z"
                                            },
                                            "verification_status": {
                                                "type": "string",
                                                "example": "verified"
                                            },
                                            "notes": {
                                                "type": "string",
                                                "example": "Account created successfully"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Email submission successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Email account submitted successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "registration_id": {
                                                    "type": "integer",
                                                    "example": 123
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "testuser123@gmail.com"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "success"
                                                },
                                                "registration_time": {
                                                    "type": "integer",
                                                    "example": 3600
                                                },
                                                "submitted_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Quota exceeded",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Monthly quota exceeded"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}